# nodejs-server/api/openapi.core.yaml
openapi: 3.0.3
info:
  title: Argovis API
  description: "REST API for Argo profiles, platforms, selections, BGC data, gridded\
    \ products, etc."
  contact:
    name: Argovis Collaboration
    url: https://argovis.colorado.edu
    email: argovis@colorado.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "2.0"
...
servers:
- url: https://argovis-api.colorado.edu
- url: /
security:
- name: []
...
paths:
...
  /cchdo:
    get:
      tags:
      - cchdo
      summary: CCHDO search and filter.
      operationId: findCCHDO
      parameters:
      - name: id
        in: query
        description: Unique ID to search for.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: 4902911_0
      - name: startDate
        in: query
        description: ISO 8601 UTC date-time formatted string indicating the beginning
          of the time period of interest.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          format: date-time
          example: 2010-01-01T00:00:00Z
      - name: endDate
        in: query
        description: ISO 8601 UTC date-time formatted string indicating the end of
          the time period of interest.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          format: date-time
          example: 2010-01-01T00:00:00Z
      - name: polygon
        in: query
        description: "array of [lon, lat] vertices describing a polygon bounding the\
          \ region of interest; final point must match initial point"
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          example: "[[-74.1,40.5],[-73.7,40.5],[-74.0,40.8],[-74.1,40.5]]"
      - name: multipolygon
        in: query
        description: array of polygon regions; region of interest is taken as the
          intersection of all listed polygons.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          example: "[[[0,0],[0,1],[1,1],[1,0],[0,0]],[[0.5,0],[0.5,1],[1.5,1],[1.5,0],[0.5,0]]]"
      - name: winding
        in: query
        description: Enforce ccw winding for polygon and multipolygon
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - "true"
      - name: center
        in: query
        description: center to measure max radius from when defining circular region
          of interest; must be used in conjunction with query string parameter 'radius'.
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "10,20.1"
          items:
            maxItems: 2
            minItems: 2
            type: number
      - name: radius
        in: query
        description: km from centerpoint when defining circular region of interest;
          must be used in conjunction with query string parameter 'center'.
        required: false
        style: form
        explode: true
        schema:
          type: number
          example: 50
      - name: metadata
        in: query
        description: metadata pointer
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: woceline
        in: query
        description: WOCE line to search for. See /cchdo/vocabulary?parameter=woceline
          for list of options.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: A10
      - name: cchdo_cruise
        in: query
        description: CCHDO cruise ID to search for. See /cchdo/vocabulary?parameter=cchdo_cruise
          for list of options.
        required: false
        style: form
        explode: true
        schema:
          type: number
          example: 518
      - name: source
        in: query
        description: Experimental program source(s) to search for; document must match
          all sources to be returned. Accepts ~ negation to filter out documents.
          See /<data route>/vocabulary?parameter=source for list of options.
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "argo_bgc,~argo_core"
          items:
            type: string
      - name: compression
        in: query
        description: Data minification strategy to apply.
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - minimal
      - name: mostrecent
        in: query
        description: get back only the n records with the most recent values of timestamp.
        required: false
        style: form
        explode: true
        schema:
          type: number
      - name: data
        in: query
        description: "Keys of data to include. Return only documents that have all\
          \ data requested, within the pressure range if specified. Accepts ~ negation\
          \ to filter out documents including the specified data. Omission of this\
          \ parameter will result in metadata only responses."
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "temperature_btl,~salinity_ctd"
          items:
            pattern: "^((~)?((ammonium)|(ammonium_error)|(ammonium_woceqc)|(arabinose)|(arabinose_woceqc)|(argon)|(argon_error)|(argon_woceqc)|(bacterial_cell_count)|(bacterial_cell_count_woceqc)|(barium)|(barium_l)|(barium_l_woceqc)|(barium_woceqc)|(bionbr)|(bnlid)|(bottle_latitude)|(bottle_longitude)|(bottle_number)|(bottle_number_woceqc)|(bottle_salinity)|(bottle_salinity_woceqc)|(bottle_time)|(brdu_uptake)|(brdu_uptake_woceqc)|(carbon_tetrachloride)|(carbon_tetrachloride_l)|(carbon_tetrachloride_l_woceqc)|(carbon_tetrachloride_woceqc)|(cdom_wavelength_300)|(cdom_wavelength_300_woceqc)|(cdom_wavelength_325)|(cdom_wavelength_325_woceqc)|(cdom_wavelength_340)|(cdom_wavelength_340_woceqc)|(cdom_wavelength_380)|(cdom_wavelength_380_woceqc)|(cdom_wavelength_412)|(cdom_wavelength_412_woceqc)|(cdom_wavelength_443)|(cdom_wavelength_443_woceqc)|(cdom_wavelength_490)|(cdom_wavelength_490_woceqc)|(cdom_wavelength_555)|(cdom_wavelength_555_woceqc)|(cdomsl)|(cdomsl_woceqc)|(cdomsn)|(cdomsn_woceqc)|(cesium_134_bq)|(cesium_134_bq_woceqc)|(cesium_137_bq)|(cesium_137_bq_error)|(cesium_137_bq_woceqc)|(cfc_11)|(cfc_113)|(cfc_113_l)|(cfc_113_l_woceqc)|(cfc_113_woceqc)|(cfc_11_error)|(cfc_11_l)|(cfc_11_l_woceqc)|(cfc_11_woceqc)|(cfc_12)|(cfc_12_error)|(cfc_12_l)|(cfc_12_l_woceqc)|(cfc_12_woceqc)|(chlorodifluoroethane)|(chlorodifluoroethane_woceqc)|(chlorodifluoromethane)|(chlorodifluoromethane_woceqc)|(chlorophyll_a)|(chlorophyll_a_ug_kg)|(chlorophyll_a_ug_kg_woceqc)|(chlorophyll_a_woceqc)|(co2_mole_fraction)|(co2_mole_fraction_woceqc)|(ctd_beamcp)|(ctd_beamcp_woceqc)|(ctd_beta650_124)|(ctd_beta700_raw)|(ctd_beta700_raw_woceqc)|(ctd_cdom)|(ctd_cdom_raw)|(ctd_cdom_raw_woceqc)|(ctd_cdom_woceqc)|(ctd_fluor)|(ctd_fluor_arbitrary)|(ctd_fluor_raw)|(ctd_fluor_raw_woceqc)|(ctd_fluor_woceqc)|(ctd_nitrate)|(ctd_nitrate_woceqc)|(ctd_number_of_observations)|(ctd_optode_oxygen)|(ctd_optode_oxygen_raw)|(ctd_optode_oxygen_raw_woceqc)|(ctd_optode_oxygen_woceqc)|(ctd_oxygen_umol_l)|(ctd_pressure_raw)|(ctd_temperature_unk)|(ctd_temperature_unk_woceqc)|(ctd_transmissometer)|(ctd_transmissometer_raw)|(ctd_transmissometer_raw_woceqc)|(ctd_transmissometer_woceqc)|(ctd_turbidity_ftu)|(ctd_turbidity_ftu_woceqc)|(ctd_turbidity_ntu)|(ctd_turbidity_ntu_woceqc)|(cyanobacteria_cell_count)|(cyanobacteria_cell_count_woceqc)|(d15n_nitrite_nitrate)|(d15n_nitrite_nitrate_woceqc)|(d15n_no3)|(d15n_no3_woceqc)|(d18o_nitrate)|(d18o_nitrate_woceqc)|(d18o_nitrite_nitrate)|(d18o_nitrite_nitrate_woceqc)|(dcns)|(dcns_woceqc)|(del_carbon_13_dic)|(del_carbon_13_dic_error)|(del_carbon_13_dic_woceqc)|(del_carbon_14_dic)|(del_carbon_14_dic_error)|(del_carbon_14_dic_woceqc)|(del_deuterium)|(del_deuterium_error)|(del_deuterium_woceqc)|(del_nitrogen_15)|(del_nitrogen_15_woceqc)|(del_oxygen_18)|(del_oxygen_18_error)|(del_oxygen_18_woceqc)|(delta_helium_3)|(delta_helium_3_error)|(delta_helium_3_woceqc)|(density_salinity)|(density_salinity_woceqc)|(dichlorofluoroethane)|(dichlorofluoroethane_woceqc)|(dissolved_organic_carbon)|(dissolved_organic_carbon_13)|(dissolved_organic_carbon_13_error)|(dissolved_organic_carbon_13_woceqc)|(dissolved_organic_carbon_14)|(dissolved_organic_carbon_14_error)|(dissolved_organic_carbon_14_woceqc)|(dissolved_organic_carbon_woceqc)|(dissolved_organic_nitrogen)|(dissolved_organic_nitrogen_woceqc)|(downcast_oxygen)|(downcast_oxygen_woceqc)|(downcast_pressure)|(downcast_pressure_woceqc)|(doxy)|(doxy_bfile)|(doxy_bfile_woceqc)|(doxy_woceqc)|(fco2)|(fco2_in_situ)|(fco2_temperature)|(fco2_woceqc)|(fdom)|(fdom_woceqc)|(fm_depth)|(fm_depth_woceqc)|(fucose)|(fucose_woceqc)|(galactose)|(galactose_woceqc)|(geotraces_sample)|(glucose)|(glucose_woceqc)|(helium)|(helium_error)|(helium_woceqc)|(hplc_19_hexanoyloxyfucoxanthin)|(hplc_19_hexanoyloxyfucoxanthin_woceqc)|(hplc_19butanoyloxyfucoxanthin)|(hplc_19butanoyloxyfucoxanthin_woceqc)|(hplc_alloxanthin)|(hplc_alloxanthin_woceqc)|(hplc_alpha_beta_carotenes)|(hplc_alpha_beta_carotenes_woceqc)|(hplc_chlorophyll_c1_c2)|(hplc_chlorophyll_c1_c2_woceqc)|(hplc_chlorophyll_c2)|(hplc_chlorophyll_c2_woceqc)|(hplc_chlorophyll_c3)|(hplc_chlorophyll_c3_woceqc)|(hplc_chlorophyllide_a)|(hplc_chlorophyllide_a_woceqc)|(hplc_diadinoxanthin)|(hplc_diadinoxanthin_woceqc)|(hplc_diatoxanthin)|(hplc_diatoxanthin_woceqc)|(hplc_divinyl_chlorophyll_a)|(hplc_divinyl_chlorophyll_a_woceqc)|(hplc_lutein)|(hplc_lutein_woceqc)|(hplc_monovinyl_chlorophyll_a)|(hplc_monovinyl_chlorophyll_a_woceqc)|(hplc_monovinyl_chlorophyll_b)|(hplc_monovinyl_chlorophyll_b_woceqc)|(hplc_neoxanthin)|(hplc_neoxanthin_woceqc)|(hplc_peridinin)|(hplc_peridinin_woceqc)|(hplc_pheophorbide_a)|(hplc_pheophorbide_a_woceqc)|(hplc_pheophytin_a)|(hplc_pheophytin_a_woceqc)|(hplc_prasinoxanthin)|(hplc_prasinoxanthin_woceqc)|(hplc_tot_chl_a)|(hplc_tot_chl_a_woceqc)|(hplc_tot_chl_b)|(hplc_tot_chl_b_woceqc)|(hplc_tot_chl_c)|(hplc_tot_chl_c_woceqc)|(hplc_violaxanthin)|(hplc_violaxanthin_woceqc)|(hplc_zeaxanthin)|(hplc_zeaxanthin_woceqc)|(hpld_antheraxanthin)|(hpld_antheraxanthin_woceqc)|(krypton)|(krypton_error)|(krypton_woceqc)|(ldeo_sample_number)|(mannose)|(mannose_woceqc)|(methane)|(methane_l)|(methane_woceqc)|(methyl_bromide)|(methyl_bromide_woceqc)|(methyl_chloride)|(methyl_chloride_woceqc)|(methyl_chloroform)|(methyl_chloroform_woceqc)|(methyl_iodide)|(methyl_iodide_woceqc)|(neon)|(neon_error)|(neon_woceqc)|(nitrate)|(nitrate_error)|(nitrate_woceqc)|(nitrite)|(nitrite_error)|(nitrite_nitrate)|(nitrite_nitrate_l)|(nitrite_nitrate_l_woceqc)|(nitrite_nitrate_woceqc)|(nitrite_woceqc)|(nitrous_oxide)|(nitrous_oxide_woceqc)|(odf_pressure)|(oxygen)|(oxygen_ml_l)|(oxygen_ml_l_woceqc)|(oxygen_woceqc)|(package_depth)|(package_depth_woceqc)|(par)|(par_raw)|(par_raw_woceqc)|(par_woceqc)|(partial_co2_temperature)|(partial_pressure_of_co2)|(partial_pressure_of_co2_woceqc)|(particulate_organic_carbon)|(particulate_organic_carbon_woceqc)|(particulate_organic_nitrogen)|(particulate_organic_nitrogen_mol)|(particulate_organic_nitrogen_mol_woceqc)|(particulate_organic_nitrogen_woceqc)|(ph_nbs)|(ph_nbs_woceqc)|(ph_sws)|(ph_sws_woceqc)|(ph_temperature)|(ph_temperature_woceqc)|(ph_total_h_scale)|(ph_total_h_scale_woceqc)|(ph_unknown_scale)|(ph_unknown_scale_woceqc)|(phaeophytin)|(phaeophytin_ug_l)|(phaeophytin_ug_l_woceqc)|(phaeophytin_woceqc)|(phosphate)|(phosphate_error)|(phosphate_l)|(phosphate_l_woceqc)|(phosphate_woceqc)|(phytoplankton_cell_count)|(phytoplankton_cell_count_woceqc)|(picoeukaryote_cell_counts)|(picoeukaryote_cell_counts_woceqc)|(potential_temperature)|(potential_temperature_woceqc)|(pressure)|(pressure_woceqc)|(prochlorophyte_cell_count)|(prochlorophyte_cell_count_woceqc)|(radium_226)|(radium_226_error)|(radium_226_woceqc)|(radium_228)|(radium_228_226)|(radium_228_226_error)|(radium_228_226_woceqc)|(radium_228_error)|(radium_228_woceqc)|(ref_temperature)|(ref_temperature_c)|(ref_temperature_c_woceqc)|(ref_temperature_woceqc)|(reference_salinity)|(reference_salinity_woceqc)|(refractive_index_anomaly)|(refractive_index_anomaly_woceqc)|(rev_pressure)|(rev_pressure_woceqc)|(rev_temperature)|(rev_temperature_90)|(rev_temperature_90_woceqc)|(rev_temperature_c)|(rev_temperature_c_woceqc)|(rev_temperature_woceqc)|(rhamnose)|(rhamnose_woceqc)|(salinity)|(salinity_bfile)|(salinity_bfile_woceqc)|(salinity_woceqc)|(sample)|(sigma0)|(sigma0_woceqc)|(silicate)|(silicate_error)|(silicate_l)|(silicate_l_woceqc)|(silicate_woceqc)|(sm_depth)|(sm_depth_woceqc)|(somma_salinity)|(spar_raw)|(spar_raw_woceqc)|(sulfur_hexifluoride)|(sulfur_hexifluoride_woceqc)|(synechococcus_cell_count)|(synechococcus_cell_count_woceqc)|(temperature)|(temperature_bfile)|(temperature_bfile_woceqc)|(temperature_woceqc)|(total_alkalinity)|(total_alkalinity_woceqc)|(total_carbon)|(total_carbon_woceqc)|(total_dissolved_nitrogen)|(total_dissolved_nitrogen_woceqc)|(total_dissolved_phosphorus)|(total_dissolved_phosphorus_woceqc)|(total_organic_carbon)|(total_organic_carbon_l)|(total_organic_carbon_l_woceqc)|(total_organic_carbon_woceqc)|(total_organic_nitrogen)|(total_organic_nitrogen_l)|(total_organic_nitrogen_l_woceqc)|(total_organic_nitrogen_woceqc)|(trifluoromethylsulfur_pentafluoride)|(trifluoromethylsulfur_pentafluoride_woceqc)|(tritium)|(tritium_error)|(tritium_woceqc)|(urea)|(urea_woceqc)|(user_bottle_number)|(user_bottle_number_woceqc)|(user_sample_number)|(user_station_number)|(xenon)|(xenon_error)|(xenon_woceqc){1})|all|except-data-values|[0-9]+)$"
            type: string
      - name: presRange
        in: query
        description: Pressure range in dbar to filter for; levels outside this range
          will not be returned.
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "10,20.1"
          items:
            maxItems: 2
            minItems: 2
            type: number
      - name: batchmeta
        in: query
        description: return the metadata documents corresponding to a temporospatial
          data search
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/cchdo'
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Cchdo
  /cchdo/meta:
    get:
      tags:
      - cchdo
      summary: GO-SHIP metadata search and filter.
      operationId: findCCHDOmeta
      parameters:
      - name: id
        in: query
        description: Unique ID to search for.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: 4902911_0
      - name: woceline
        in: query
        description: WOCE line to search for. See /cchdo/vocabulary?parameter=woceline
          for list of options.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: A10
      - name: cchdo_cruise
        in: query
        description: CCHDO cruise ID to search for. See /cchdo/vocabulary?parameter=cchdo_cruise
          for list of options.
        required: false
        style: form
        explode: true
        schema:
          type: number
          example: 518
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/cchdoMeta'
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Cchdo
  /cchdo/vocabulary:
    get:
      tags:
      - cchdo
      summary: List all possible values for certain CCHDO query string parameters
      operationId: cchdoVocab
      parameters:
      - name: parameter
        in: query
        description: GO-SHIP query string parameter to summarize possible values of.
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - woceline
          - cchdo_cruise
          - source
          - data
          - metadata
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Cchdo
...
components:
  schemas:
...
    geoJSONpoint:
      required:
      - coordinates
      - type
      type: object
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            maxItems: 2
            minItems: 2
            type: number
      example:
        coordinates:
        - 0.80082819046101150206595775671303272247314453125
        - 0.80082819046101150206595775671303272247314453125
        type: type
...
    data:
      type: array
      items:
        type: array
        items:
          anyOf:
          - type: string
          - type: number
    data_info:
      type: array
      items:
        anyOf:
        - type: array
          items:
            type: string
        - type: array
          items:
            type: array
            items:
              anyOf:
              - type: string
              - type: number
    generic_source_info:
      type: object
      properties:
        source:
          type: array
          items:
            type: string
        url:
          type: string
        date_updated:
          type: string
          format: date-time
        doi:
          type: string
      example:
        date_updated: 2000-01-23T04:56:07.000+00:00
        source:
        - source
        - source
        url: url
        doi: doi
...
    cchdo:
      required:
      - _id
      - basin
      - cast
      - data_info
      - geolocation
      - metadata
      - source
      - station
      - timestamp
      type: object
      properties:
        _id:
          type: string
        metadata:
          type: array
          items:
            type: string
        geolocation:
          $ref: '#/components/schemas/geoJSONpoint'
        btm_depth:
          type: number
        file_hash:
          type: string
        basin:
          type: number
        timestamp:
          type: string
          format: date-time
        source:
          type: array
          items:
            $ref: '#/components/schemas/generic_source_info'
        data_warning:
          type: array
          items:
            type: string
            enum:
            - degenerate_levels
            - missing_basin
            - missing_location
            - missing_timestamp
        station:
          type: string
        cast:
          type: number
        data:
          $ref: '#/components/schemas/data'
        data_info:
          $ref: '#/components/schemas/data_info'
      example:
        metadata:
        - metadata
        - metadata
        data:
        - - ""
          - ""
        - - ""
          - ""
        basin: 6.02745618307040320615897144307382404804229736328125
        source:
        - date_updated: 2000-01-23T04:56:07.000+00:00
          source:
          - source
          - source
          url: url
          doi: doi
        - date_updated: 2000-01-23T04:56:07.000+00:00
          source:
          - source
          - source
          url: url
          doi: doi
        btm_depth: 0.80082819046101150206595775671303272247314453125
        data_warning:
        - degenerate_levels
        - degenerate_levels
        cast: 1.46581298050294517310021547018550336360931396484375
        data_info:
        - ""
        - ""
        file_hash: file_hash
        station: station
        _id: _id
        geolocation:
          coordinates:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          type: type
        timestamp: 2000-01-23T04:56:07.000+00:00
    cchdoMeta:
      required:
      - _id
      - cchdo_cruise_id
      - data_type
      - date_updated_argovis
      - expocode
      - woce_lines
      type: object
      properties:
        _id:
          type: string
        date_updated_argovis:
          type: string
          format: date-time
        data_type:
          type: string
        country:
          type: string
        data_center:
          type: string
        instrument:
          type: string
        pi_name:
          type: array
          items:
            type: string
        expocode:
          type: string
        file_expocode:
          type: string
        cchdo_cruise_id:
          type: number
        woce_lines:
          type: array
          items:
            type: string
        positioning_system:
          type: string
      example:
        country: country
        positioning_system: positioning_system
        expocode: expocode
        woce_lines:
        - woce_lines
        - woce_lines
        pi_name:
        - pi_name
        - pi_name
        file_expocode: file_expocode
        cchdo_cruise_id: 0.80082819046101150206595775671303272247314453125
        data_type: data_type
        data_center: data_center
        instrument: instrument
        _id: _id
        date_updated_argovis: 2000-01-23T04:56:07.000+00:00
...
    errorResponse:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...